api/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0,<2.0.0
pydantic[email]>=1.8.0,<2.0.0
python-dotenv>=0.19.0,<0.20.0
pytest>=6.2.5,<7.0.0
requests>=2.26.0,<3.0.0

# Optional: PostgreSQL support
# psycopg2-binary>=2.9.0,<3.0.0 